qa_app/requirements.txt
flask==3.0.2
opensearch-py==2.4.2
python-dotenv==1.0.1 